refactor(test): create ERPNext specific test suite
and move data initialization there
